pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "security-tools"
version = "0.2.0"
description = "Security review, runtime verification, and artifact-centric DevSecOps control plane toolkit"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "pydantic>=2.8,<3",
  "pydantic-settings>=2.4,<3",
  "requests>=2.32,<3",
  "python-dotenv>=1.0,<2",
  "PyYAML>=6.0,<7",
  "PyMuPDF>=1.24,<2",
]

[project.scripts]
security-review = "security_tools.cli:main"
container-runtime-verify = "security_tools.runtime.cli:main"
control-plane-evaluate = "security_tools.control_plane.commands.evaluate_promotion:main"
control-plane-promote = "security_tools.control_plane.commands.promote_artifact:main"
control-plane-record-deployment = "security_tools.control_plane.commands.record_deployment:main"
control-plane-check-drift = "security_tools.control_plane.commands.check_drift:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["security_tools*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
"security_tools.catalog" = ["*.yml", "*.yaml"]
"security_tools.intelligence.knowledge" = ["**/*.yml", "**/*.yaml", "*.yml", "*.yaml"]
"security_tools.control_plane.config" = ["*.yml", "*.yaml"]